Change issue_number datatype to float
Some series occasionally have a second storyline running in tandem with the primary book and use #issueNumber.1 numbering convention (e.g., https://www.marvel.com/comics/issue/51327/the_amazing_spider-man_2014_16.1).  This makes int datatype unviable for storing issue numbers.

+++ models/comic.js
@@ -21,7 +21,7 @@ module.exports = (sequelize, DataTypes) => {
       marvel_id: DataTypes.INTEGER,
       title: DataTypes.STRING,
       series: DataTypes.STRING,
-      issue_number: DataTypes.INTEGER,
+      issue_number: DataTypes.FLOAT,
       thumbnail_url: DataTypes.STRING,
       marvel_url: DataTypes.STRING,
       user_id: DataTypes.INTEGER,
